The big dumb refactor
