+++ worlds/witness/data/WitnessLogicExpert.txt
