+++ worlds/witness/data/settings/Door_Shuffle/Complex_Additional_Panels.txt
