+++ worlds/witness/data/settings/Door_Shuffle/Elevators_Come_To_You.txt
