+++ worlds/witness/data/settings/EP_Shuffle/EP_Easy.txt
