+++ worlds/witness/data/settings/EP_Shuffle/EP_Sides.txt
