+++ worlds/witness/data/settings/Exclusions/Discards.txt
