+++ worlds/witness/data/settings/Postgame/Bottom_Floor_Discard.txt
