+++ worlds/witness/data/static_logic.py
@@ -1,56 +1,14 @@
-from dataclasses import dataclass
-from enum import Enum
+from functools import lru_cache
 from typing import Dict, List
 
-from .utils import define_new_region, parse_lambda, lazy, get_items, get_sigma_normal_logic, get_sigma_expert_logic,\
+from .item_definition_classes import ItemCategory, CATEGORY_NAME_MAPPINGS, DoorItemDefinition, WeightedItemDefinition, \
+    ProgressiveItemDefinition, ItemDefinition
+from worlds.witness.utils import define_new_region, parse_lambda, get_items, get_sigma_normal_logic, get_sigma_expert_logic,\
     get_vanilla_logic
 
 
-class ItemCategory(Enum):
-    SYMBOL = 0
-    DOOR = 1
-    LASER = 2
-    USEFUL = 3
-    FILLER = 4
-    TRAP = 5
-    JOKE = 6
-    EVENT = 7
-
-
-CATEGORY_NAME_MAPPINGS: Dict[str, ItemCategory] = {
-    "Symbols:": ItemCategory.SYMBOL,
-    "Doors:": ItemCategory.DOOR,
-    "Lasers:": ItemCategory.LASER,
-    "Useful:": ItemCategory.USEFUL,
-    "Filler:": ItemCategory.FILLER,
-    "Traps:": ItemCategory.TRAP,
-    "Jokes:": ItemCategory.JOKE
-}
-
-
-@dataclass(frozen=True)
-class ItemDefinition:
-    local_code: int
-    category: ItemCategory
-
-
-@dataclass(frozen=True)
-class ProgressiveItemDefinition(ItemDefinition):
-    child_item_names: List[str]
-
-
-@dataclass(frozen=True)
-class DoorItemDefinition(ItemDefinition):
-    panel_id_hexes: List[str]
-
-
-@dataclass(frozen=True)
-class WeightedItemDefinition(ItemDefinition):
-    weight: int
-
-
 class StaticWitnessLogicObj:
-    def read_logic_file(self, lines):
+    def read_logic_file(self, lines) -> None:
         """
         Reads the logic file and does the initial population of data structures
         """
@@ -177,7 +135,7 @@ class StaticWitnessLogicObj:
 
             current_region["panels"].append(entity_hex)
 
-    def __init__(self, lines=None):
+    def __init__(self, lines=None) -> None:
         if lines is None:
             lines = get_sigma_normal_logic()
 
@@ -199,102 +157,95 @@ class StaticWitnessLogicObj:
         self.read_logic_file(lines)
 
 
-class StaticWitnessLogic:
-    # Item data parsed from WitnessItems.txt
-    all_items: Dict[str, ItemDefinition] = {}
-    _progressive_lookup: Dict[str, str] = {}
-
-    ALL_REGIONS_BY_NAME = dict()
-    ALL_AREAS_BY_NAME = dict()
-    STATIC_CONNECTIONS_BY_REGION_NAME = dict()
-
-    OBELISK_SIDE_ID_TO_EP_HEXES = dict()
-
-    ENTITIES_BY_HEX = dict()
-    ENTITIES_BY_NAME = dict()
-    STATIC_DEPENDENT_REQUIREMENTS_BY_HEX = dict()
-
-    EP_TO_OBELISK_SIDE = dict()
-
-    ENTITY_ID_TO_NAME = dict()
-
-    @staticmethod
-    def parse_items():
-        """
-        Parses currently defined items from WitnessItems.txt
-        """
-
-        lines: List[str] = get_items()
-        current_category: ItemCategory = ItemCategory.SYMBOL
-
-        for line in lines:
-            # Skip empty lines and comments.
-            if line == "" or line[0] == "#":
-                continue
-
-            # If this line is a category header, update our cached category.
-            if line in CATEGORY_NAME_MAPPINGS.keys():
-                current_category = CATEGORY_NAME_MAPPINGS[line]
-                continue
-
-            line_split = line.split(" - ")
-
-            item_code = int(line_split[0])
-            item_name = line_split[1]
-            arguments: List[str] = line_split[2].split(",") if len(line_split) >= 3 else []
-
-            if current_category in [ItemCategory.DOOR, ItemCategory.LASER]:
-                # Map doors to IDs.
-                StaticWitnessLogic.all_items[item_name] = DoorItemDefinition(item_code, current_category,
-                                                                             arguments)
-            elif current_category == ItemCategory.TRAP or current_category == ItemCategory.FILLER:
-                # Read filler weights.
-                weight = int(arguments[0]) if len(arguments) >= 1 else 1
-                StaticWitnessLogic.all_items[item_name] = WeightedItemDefinition(item_code, current_category, weight)
-            elif arguments:
-                # Progressive items.
-                StaticWitnessLogic.all_items[item_name] = ProgressiveItemDefinition(item_code, current_category,
-                                                                                    arguments)
-                for child_item in arguments:
-                    StaticWitnessLogic._progressive_lookup[child_item] = item_name
-            else:
-                StaticWitnessLogic.all_items[item_name] = ItemDefinition(item_code, current_category)
-
-    @staticmethod
-    def get_parent_progressive_item(item_name: str):
-        """
-        Returns the name of the item's progressive parent, if there is one, or the item's name if not.
-        """
-        return StaticWitnessLogic._progressive_lookup.get(item_name, item_name)
-
-    @lazy
-    def sigma_expert(self) -> StaticWitnessLogicObj:
-        return StaticWitnessLogicObj(get_sigma_expert_logic())
-
-    @lazy
-    def sigma_normal(self) -> StaticWitnessLogicObj:
-        return StaticWitnessLogicObj(get_sigma_normal_logic())
-
-    @lazy
-    def vanilla(self) -> StaticWitnessLogicObj:
-        return StaticWitnessLogicObj(get_vanilla_logic())
-
-    def __init__(self):
-        self.parse_items()
-
-        self.ALL_REGIONS_BY_NAME.update(self.sigma_normal.ALL_REGIONS_BY_NAME)
-        self.ALL_AREAS_BY_NAME.update(self.sigma_normal.ALL_AREAS_BY_NAME)
-        self.STATIC_CONNECTIONS_BY_REGION_NAME.update(self.sigma_normal.STATIC_CONNECTIONS_BY_REGION_NAME)
-
-        self.ENTITIES_BY_HEX.update(self.sigma_normal.ENTITIES_BY_HEX)
-        self.ENTITIES_BY_NAME.update(self.sigma_normal.ENTITIES_BY_NAME)
-        self.STATIC_DEPENDENT_REQUIREMENTS_BY_HEX.update(self.sigma_normal.STATIC_DEPENDENT_REQUIREMENTS_BY_HEX)
-
-        self.OBELISK_SIDE_ID_TO_EP_HEXES.update(self.sigma_normal.OBELISK_SIDE_ID_TO_EP_HEXES)
-
-        self.EP_TO_OBELISK_SIDE.update(self.sigma_normal.EP_TO_OBELISK_SIDE)
-
-        self.ENTITY_ID_TO_NAME.update(self.sigma_normal.ENTITY_ID_TO_NAME)
+# Item data parsed from WitnessItems.txt
+ALL_ITEMS: Dict[str, ItemDefinition] = {}
+_progressive_lookup: Dict[str, str] = {}
 
 
-StaticWitnessLogic()
+def parse_items() -> None:
+    """
+    Parses currently defined items from WitnessItems.txt
+    """
+
+    lines: List[str] = get_items()
+    current_category: ItemCategory = ItemCategory.SYMBOL
+
+    for line in lines:
+        # Skip empty lines and comments.
+        if line == "" or line[0] == "#":
+            continue
+
+        # If this line is a category header, update our cached category.
+        if line in CATEGORY_NAME_MAPPINGS.keys():
+            current_category = CATEGORY_NAME_MAPPINGS[line]
+            continue
+
+        line_split = line.split(" - ")
+
+        item_code = int(line_split[0])
+        item_name = line_split[1]
+        arguments: List[str] = line_split[2].split(",") if len(line_split) >= 3 else []
+
+        if current_category in [ItemCategory.DOOR, ItemCategory.LASER]:
+            # Map doors to IDs.
+            ALL_ITEMS[item_name] = DoorItemDefinition(item_code, current_category, arguments)
+        elif current_category == ItemCategory.TRAP or current_category == ItemCategory.FILLER:
+            # Read filler weights.
+            weight = int(arguments[0]) if len(arguments) >= 1 else 1
+            ALL_ITEMS[item_name] = WeightedItemDefinition(item_code, current_category, weight)
+        elif arguments:
+            # Progressive items.
+            ALL_ITEMS[item_name] = ProgressiveItemDefinition(item_code, current_category, arguments)
+            for child_item in arguments:
+                _progressive_lookup[child_item] = item_name
+        else:
+            ALL_ITEMS[item_name] = ItemDefinition(item_code, current_category)
+
+
+def get_parent_progressive_item(item_name: str) -> str:
+    """
+    Returns the name of the item's progressive parent, if there is one, or the item's name if not.
+    """
+    return _progressive_lookup.get(item_name, item_name)
+
+
+@lru_cache
+def get_vanilla() -> StaticWitnessLogicObj:
+    return StaticWitnessLogicObj(get_sigma_expert_logic())
+
+
+@lru_cache
+def get_sigma_normal() -> StaticWitnessLogicObj:
+    return StaticWitnessLogicObj(get_sigma_normal_logic())
+
+
+@lru_cache
+def get_sigma_expert() -> StaticWitnessLogicObj:
+    return StaticWitnessLogicObj(get_vanilla_logic())
+
+
+def __getattr__(name):
+    if name == "vanilla":
+        return get_vanilla()
+    elif name == "sigma_normal":
+        return get_sigma_normal()
+    elif name == "sigma_expert":
+        return get_sigma_expert()
+    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
+
+
+parse_items()
+
+ALL_REGIONS_BY_NAME = get_sigma_normal().ALL_REGIONS_BY_NAME
+ALL_AREAS_BY_NAME = get_sigma_normal().ALL_AREAS_BY_NAME
+STATIC_CONNECTIONS_BY_REGION_NAME = get_sigma_normal().STATIC_CONNECTIONS_BY_REGION_NAME
+
+ENTITIES_BY_HEX = get_sigma_normal().ENTITIES_BY_HEX
+ENTITIES_BY_NAME = get_sigma_normal().ENTITIES_BY_NAME
+STATIC_DEPENDENT_REQUIREMENTS_BY_HEX = get_sigma_normal().STATIC_DEPENDENT_REQUIREMENTS_BY_HEX
+
+OBELISK_SIDE_ID_TO_EP_HEXES = get_sigma_normal().OBELISK_SIDE_ID_TO_EP_HEXES
+
+EP_TO_OBELISK_SIDE = get_sigma_normal().EP_TO_OBELISK_SIDE
+
+ENTITY_ID_TO_NAME = get_sigma_normal().ENTITY_ID_TO_NAME
